# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Blood Donation System API",
    "version": "0.1.0"
  },
  "paths": {
    "/donors/": {
      "get": {
        "tags": [
          "donors"
        ],
        "summary": "Get Donors",
        "operationId": "get_donors_donors__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DonorRead"
                  },
                  "type": "array",
                  "title": "Response Get Donors Donors  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "donors"
        ],
        "summary": "Create Donor",
        "operationId": "create_donor_donors__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DonorCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DonorRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/donors/{donor_id}": {
      "get": {
        "tags": [
          "donors"
        ],
        "summary": "Get Donor",
        "operationId": "get_donor_donors__donor_id__get",
        "parameters": [
          {
            "name": "donor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Donor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DonorRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "donors"
        ],
        "summary": "Delete Donor",
        "operationId": "delete_donor_donors__donor_id__delete",
        "parameters": [
          {
            "name": "donor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Donor Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Read Root",
        "operationId": "read_root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BloodType": {
        "type": "string",
        "enum": [
          "A+",
          "A-",
          "B+",
          "B-",
          "AB+",
          "AB-",
          "O+",
          "O-"
        ],
        "title": "BloodType"
      },
      "DonorCreate": {
        "properties": {
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "blood_type": {
            "$ref": "#/components/schemas/BloodType"
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "email",
          "blood_type"
        ],
        "title": "DonorCreate"
      },
      "DonorRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "blood_type": {
            "$ref": "#/components/schemas/BloodType"
          }
        },
        "type": "object",
        "required": [
          "id",
          "full_name",
          "email",
          "blood_type"
        ],
        "title": "DonorRead"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}